requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
Django==4.2.4
django-redis==5.3.0
djangorestframework==3.14.0
httpie==3.2.2
idna==3.4
kombu==5.3.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
prompt-toolkit==3.0.39
Pygments==2.16.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==3.5.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
python-decouple
